fix: spinner race condition
There are two things writing to stderr, spinner and error messages.
Chat writes to stdout.

Run Chat in a goroutine and wait for it to finish. Meanwhile, run the
spinner with stderr in main. Once Chat finishes, exit the spinner and
print output to stdout.

This also uses a separate Lipgloss renderer for stderr instead of doing
the hacky `lipgloss.SetColorProfile(termenv.NewOutput...)`.

Add `prettyError` to style error messages with lipgloss.

Signed-off-by: Ayman Bagabas <ayman.bagabas@gmail.com>