feat: have --quiet supress all stderr messages for success
Before this commit, according to the README, --quiet should supress all
stderr output, which was incorrect. It did what the help says: supress
only the spinner.

This commit makes --quiet supress all "superflous" stderr ouput, meaning
both the spinner and messages that confirm an operation was carried out
successfully like "Conversation saved...".

Rationale: for scriptability, it would be nice to be able to supress
non-error stderr messages without losing the ability to see the error
when something actually goes wrong.

Maybe this isn't the way to do it though, and there should be options
for --no-spinner, --quiet (supress all stderr) and --shy (what this
commit implements).