fix: improve --continue
If you run `mods` without any args or input, it'll print the help.
If you run `mods` with either a prompt or piping something from stdin,
it uses that as a prompt.

`mods -C` was requiring a prompt even when it was provided via stdin.
This removes that check, so it works the same as `mods` without the `-C`
flag now.

closes #132