fix: lint complaining about capitalized errors
That complaint makes sense to some extend, but in our case, we want
these errors capitalized, as they are user-facing.

Created an `errors.go` file, moved `modsError` in there, and added a
`newUserErrorf` function to bypass that lint check.

PS: it is not `golangci-lint` that is throwing these errors, its
`gopls`, and I couldn't find a way to disable it per-line.

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>