feat: support json
Introduce a new flag, `--format-as`.

It defaults to `markdown`, and is only used when `--format` is also
specified.

`--format-as=json` will only set the response type to json in supported models, otherwise it'll only add the format text to the prompt.

Also updated the openai library to support this.

Added to the default configuration both `gpt-4-1106-preview` and
`gpt-3.5-turbo-1106`.

Theses are the only models that support the JSON response format.

`format-text` is now a `map[string]string`. Old configs (which have it
as a string) will set the specified formatting text to the `markdown`
format automatically.
Newly generated configs will have both `markdown` and `json` format
texts.

---

closes #167
closes #173

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>