fix: improve completion cmd handling
The way it works today, its not possible to install the completions
automatically on Nix, as the build needs to be pure.

This is kinda hacky, but I think it'll work: we only allow the
completion cmd to run if there is good evidence that it is a completion
command, and not a prompt that happens to start with completion.

refs #287