feat: Add support for user request parameter (#362)
In certain OpenAI Azure setups, the `user` parameter is mandatory in the requests. Failing to include this parameter results in a 422 error code.
This patch introduces a mechanism to add the user parameter, allowing the use of mods in such setups.
I would appreciate your feedback on this change and any suggestions for improvement.
Thank you!